Better seperation of concerns
